rename: Move utils file to client folder
- validate.ts, index.js를 client 폴더로 이동
- 이에 따라 경로 변경
